+++ librtp/payload/rtp-h265-pack.c
@@ -83,7 +83,7 @@ static int rtp_h265_pack_nalu(struct rtp_encode_h265_t *packer, const uint8_t* n
 	packer->pkt.rtp.m = 1; // set marker flag
 	//packer->pkt.rtp.m = ((*nalu >> 1) & 0x3f) < 32 ? 1 : 0;
 	n = rtp_packet_serialize(&packer->pkt, rtp, n);
-	if ((size_t)n != RTP_FIXED_HEADER + packer->pkt.payloadlen)
+	if (n != RTP_FIXED_HEADER + packer->pkt.payloadlen)
 	{
 		assert(0);
 		return -1;
@@ -127,7 +127,7 @@ static int rtp_h265_pack_fu(struct rtp_encode_h265_t *packer, const uint8_t* ptr
 
 		packer->pkt.rtp.m = (FU_END & fu_header) ? 1 : 0; // set marker flag
 		n = rtp_packet_serialize(&packer->pkt, rtp, n);
-		if ((size_t)n != RTP_FIXED_HEADER + packer->pkt.payloadlen)
+		if (n != RTP_FIXED_HEADER + packer->pkt.payloadlen)
 		{
 			assert(0);
 			return -1;
